Added some assets and folder preset scripts
I was working on an inventory system but ultimately did not like how it turned out and wanted to restart from scratch. I did however added some assets in this commit as well as a script for customizing default settings to specific folders that may be useful in the future.